Band-limit the resampler, and stop truncating callback arguments
Two unrelated things, found by chasing each other.

── Resampler ──
13379 Hz -> 44100 Hz was linear interpolation, whose sinc^2 response is a
poor anti-imaging filter.  Measured on real output with the PSG muted — so
everything above the mixer's 6.69 kHz Nyquist is necessarily an artifact —
the images sat this far below the signal:

              6.7-9k   9-13.4k   13.4-17k   17-22k
  linear      -24.0     -33.2      -45.1     -37.2  dB
  windowed    -28.5     -80.5      -81.8     -80.3  dB

Now a Kaiser-windowed sinc (beta 9, 24 input samples each side), tabulated
at 512 steps per input sample.  The taps are one input sample apart, so the
table index strides by a constant and the phase is computed once per output
sample rather than per tap.  Cost 0.27 ms/frame at -O2, 0.80 at -O0; delay
1.79 ms, because a symmetric kernel needs look-ahead the newest block does
not have, so output now runs out of a history ring.

The residual right above 6.69 kHz is the filter's own transition band.
Pushing it below 300 Hz wide would need ~250 taps and 9 ms of delay, to
suppress mirror images of content at 6.4-6.7 kHz.  Not worth it.

RTPC_AUDIO_LINEAR=1 restores the old interpolator for comparison.

── Callback-argument truncation ──
Widening the mixer's rate knob led to a -O2 experiment, which crashed
data_room in text_printer.c.  The faulting pointer was 0x00000000f023f940 —
a 64-bit pointer with the top half gone — and the source was

    text_printer_run_func_on_finish(listbox->printer, func_0800ae88, (s32)listbox);

against a `u32 onFinishArg` field.  The pattern is systemic: every
callback-argument slot in the port was s32/u32, and callers pass pointers
through them.  A scan with -Wpointer-to-int-cast (normally suppressed here)
found 34 sites; this widens the whole chain to intptr_t/uintptr_t —
TextPrinter, Listbox, sprite and affine-sprite callbacks, the task pool,
ScheduledFunctionTask, the beatscript scene callback, gameplay engine
events (whose *return* value is cast back to a SongHeader *), and the
explicit truncating casts in sick_beats, night_walk and sneaky_spirits.
CALL_DRUM_PLAY_FUNC's `|1` is the ARM/Thumb interworking bit; on PC it
truncated a pointer and then corrupted what was left, so it is now GBA-only.

Measured over six runs each, before -> after:

  data_room             6/6 crash -> 3/6
  reading               3/6 crash -> 0-1/6
  drum_studio_drumming  6/6 crash -> 6/6   (untouched by this)
  karate_man            0/6       -> 0/6

So this is a real reduction, not a fix: at least one corruption bug remains
in the text scenes.  data_room does now reach 19 colours instead of dying
during boot.  The remaining truncations are in DMA-register writes and one
GBA-only branch, none of which are read as pointers on PC.

drum_studio_drumming crashes identically before and after, deterministically:
thread->subScene comes back as 0x00000030_010d1ab4 where the top half should
be 0x00000001, so something writes 32 bits over a 64-bit field.  Recorded,
not fixed.

Makefile.pc gains OPT.  -O2 halves per-frame work (7.17 -> 3.49 ms over 1300
frames), but the default stays -O0 while the corruption above is outstanding
— -O2 neither causes nor helps it, and this is not the commit to change what
ships.

+++ data/data_08936b54.h
@@ -242,8 +242,8 @@ struct unk_struct_08008420 {
 };
 
 struct ScheduledFunctionTask {
-    void (*function)(s32);
-    s32 param;
+    void (*function)(intptr_t);
+    intptr_t param;
     u32 delay;
 };
